Respect symlinks in url_stat
url_stat is called with an additional $flags argument by PHP.
This includes an flag whether info about sylimk or symlink target should
be returned.

Prior to this change, stat was used all the time, delivering results for
the target of symlinks.
Respecting the flag, lstat is used to return info about the symlink
instead, when requested.

This will properly return infos for is_link() calls.